init: allow booting into recovery according to the cmdline
Samsung Vibrant, Galaxy S and Captivate only have a single Kernel partition
The Bootloader supplies bootmode=2 to the kernel cmdline when we should boot
into recovery. If it sees that value on cmdline we only parse the recovery.rc
in all other cases it parses the board specific rc file. We tested on DHD where
the bootloader sends androidboot.mode=recovery in case of recovery, so this
shouldn't conflict but it needs testing on other platforms.

Change-Id: I4ccc66833298e7185537ed505c60a0aa0602e814

coolya (author)